pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-context-processor"
version = "1.0.0"
description = "Convert documents to markdown using MarkItDown for LLM context"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Anonymous", email = "anonymous@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "markitdown[all]>=0.1.0",
    "pyyaml>=6.0.1",
]

[project.urls]
"Homepage" = "https://github.com/YOUR_USERNAME/llm-context-processor"
"Repository" = "https://github.com/YOUR_USERNAME/llm-context-processor"

[project.scripts]
llm-context-processor = "llm_context_processor.cli:main"

[tool.setuptools]
packages = [
    "llm_context_processor",
    "llm_context_processor.extractors",
    "llm_context_processor.outputs",
    "llm_context_processor.utils"
]
